+++ NEWS.md
@@ -5,3 +5,7 @@
 * Package initialized.
 * Added a `NEWS.md` file to track changes to the package.
 * performed devtools::check() and testthat with flying colors, ready for full scale testing. 
+
+## 2026-02-27
+* Added node option for routing (i.e. inflow = outflow) when cell area = 0 km2
+* added `dmstar_default_params` function to build a default parameter list to be used with DMSTAr simulation functions